+++ R/raw.writeRAWIDfile.R
@@ -2,15 +2,19 @@
 #'
 #' @param rID RAW ID data frame
 #' @param header_list list to be saved
-#' @param fIDfile filen ame RAW_ID.csv to be saved
+#' @param fIDfile filename RAW_ID.csv to be saved
 #'
 #'
 #' @importFrom utils write.csv write.table
 #'
 #' @export
 raw.writeRAWIDfile <- function(rID, header_list, fIDfile) {
-  write.csv(rID, file=fIDfile, row.names = FALSE)
   raw.writeRAWIDheader(header_list, fIDfile)
+  suppressWarnings(
+    write.table(rID, file=fIDfile, row.names = FALSE,
+                append=TRUE, sep=",", col.names = TRUE)
+    )
+  # write.csv(rID, file=fIDfile, row.names = FALSE)
 }
 
 #' Writes a RAW ID Header
@@ -33,6 +37,6 @@ raw.writeRAWIDheader <- function(header_list, fIDfile) {
     header_data = c(header_data, paste("#", keys, ":", v))
   }
   write.table(header_data, file=fIDfile,
-              append=TRUE, quote=FALSE,
+              append=FALSE, quote=FALSE,
               col.names = FALSE, row.names = FALSE)
 }